Refactored publishers and keywords lists - fewer files, much easier on the eyes
