--- app/components/keywordlistitem.js
@@ -1,21 +1,0 @@
-import React from 'react'
-import Fluxxor from 'fluxxor'
-var FluxMixin = Fluxxor.FluxMixin(React)
-require("font-awesome-webpack");
-
-export default React.createClass({
-  mixins: [FluxMixin],
-  handleRemoveKeyword: function(keywordIndex) {
-    this.getFlux().actions.removeKeyword(this.props.chartID, keywordIndex);
-  },
-
-  render: function() {
-    var keyword = this.props.keyword
-    return (
-      <li className="keyword-list-item">
-        {keyword} &nbsp;
-        <i className="fa fa-times-circle-o" onClick={this.handleRemoveKeyword.bind(this, this.props.id)}></i>
-      </li>
-    )
-  }
-})
